+++ Magic/CHANGELOG.md
@@ -14,6 +14,7 @@
    - Add "example: heroes" configs with craftable Heroes wand
    - Add Magic Skill Pack, all default Magic spells packaged as skills:
      https://s3-us-west-2.amazonaws.com/elmakers/Magic/MagicHeroesSkills.zip
+   - Support pass-through parameters from skill configs to Magic spell skills
 
 ## 6.6